requirements.txt
alembic==1.7.4
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Admin==1.5.8
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.8.1
importlib-resources==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
pdfkit==0.6.1   
phonenumbers==8.12.35
Pillow==8.4.0
pycountry==20.7.3
pycparser==2.20
python-dotenv==0.19.1
six==1.16.0
SQLAlchemy==1.4.26
Werkzeug==2.0.2
WTForms==2.3.3
zipp==3.6.0
